part 1 working

+++ A03/main.c
@@ -7,7 +7,7 @@
 #include "sieve_odd.h"
 
 
-void validate(unsigned int a, unsigned int b, char *msg){
+void validate(uint64_t a, uint64_t b, char *msg){
 
 	char *check = a == b ? "(PASS)" : "(FAIL)";
 	printf("Comparison between %s,%s\n",msg,check);
@@ -17,7 +17,7 @@ void validate(unsigned int a, unsigned int b, char *msg){
 int main(int argc, char **argv){
 	int id;	// My Id
 	int p;	// Number of processes
-	unsigned int n;
+	uint64_t n;
 	int exp;
 
 	MPI_Init (&argc, &argv);
@@ -34,9 +34,9 @@ int main(int argc, char **argv){
 	n = pow(10,(double)exp);
 
 
-	unsigned int res_orig=sieve_original(id,n,p);
+	uint64_t res_orig=sieve_original(id,n,p);
 	MPI_Barrier(MPI_COMM_WORLD);
-	unsigned int res_odd=sieve_odd(id,n,p);
+	uint64_t res_odd=sieve_odd(id,n,p);
 	MPI_Barrier(MPI_COMM_WORLD);
 
 	if(id==0) validate(res_orig,res_odd,"original to odd version");
