+++ A03/sieve_odd.h
@@ -6,12 +6,12 @@
 #define ODD_INDEX(x) (((x-3)/2))
 #define INDEX_ODD(x) (2*x + 3)
 
-unsigned int sieve_odd(int id, unsigned int n,unsigned int p){
+uint64_t sieve_odd(int id, uint64_t n,uint64_t p){
 	if(id == 0) printf("Executing odd numbers sieve\n");
-	unsigned int low_value = (id==0) ? 3 : 2 + id*((n-1))/p;
+	uint64_t low_value = (id==0) ? 3 : 2 + ((uint64_t)id)*((n-1))/p;
 	if(!(id==0) ) low_value = (low_value % 2 == 0) ? low_value+1 : low_value;//Start with odd
-	unsigned int high_value = 1 + (id+1)*((n-1))/p;
-	unsigned int size = (high_value - low_value)/2 + 1;
+	uint64_t high_value = 1 + ((uint64_t)(id+1))*((n-1))/p;
+	uint64_t size = (high_value - low_value)/2 + 1;
 
 	/*int k;
 	int c=-1;
@@ -23,14 +23,15 @@ unsigned int sieve_odd(int id, unsigned int n,unsigned int p){
 	}*/
 
 	//MPI_Barrier(MPI_COMM_WORLD);
-	//if(id==c)printf("([%d],%d,%d,%d)\n",id,low_value,high_value,size);
+	//if(id==c)
+	//printf("([%d],%"PRIu64",%"PRIu64",%"PRIu64")\n",id,low_value,high_value,size);
 	MPI_Barrier(MPI_COMM_WORLD);
 
 	double elapsed_time = -MPI_Wtime();
 
-	unsigned int proc0_size = (n-1)/p;
+	uint64_t proc0_size = (n-1)/p;
 
-	if ((2 + proc0_size) < (int) sqrt((double) n)) {
+	if ((2 + proc0_size) < (unsigned int) sqrt((double) n)) {
 		if (id == 0) printf ("Too many processes\n");
 		MPI_Finalize();
 		exit (1);
@@ -43,12 +44,12 @@ unsigned int sieve_odd(int id, unsigned int n,unsigned int p){
 		exit (1);
 	}
 
-	unsigned int i;
+	uint64_t i;
 	for (i = 0; i < size; i++) marked[i] = 0;
-	unsigned int index;
+	uint64_t index;
 	if (id == 0) index = 0;
-	unsigned int prime = 3;
-	unsigned int first;
+	uint64_t prime = 3;
+	uint64_t first;
 
 	do
 	{
@@ -72,7 +73,7 @@ unsigned int sieve_odd(int id, unsigned int n,unsigned int p){
 		}
 
 		i=first;
-		unsigned int offset = ODD_INDEX(low_value);
+		uint64_t offset = ODD_INDEX(low_value);
 		while( i <= high_value){
 			marked[ODD_INDEX(i) - offset]=1;
 			i+=(prime <<1);
@@ -82,7 +83,7 @@ unsigned int sieve_odd(int id, unsigned int n,unsigned int p){
 			while (marked[++index]);
 			prime = INDEX_ODD(index);
 		}
-		if( p > 1 ) MPI_Bcast (&prime,  1, MPI_INT, 0, MPI_COMM_WORLD);
+		if( p > 1 ) MPI_Bcast (&prime,  1, MPI_UNSIGNED_LONG, 0, MPI_COMM_WORLD);
 	}while(prime * prime <= n);
 
 	/*if(id==c){
@@ -93,17 +94,17 @@ unsigned int sieve_odd(int id, unsigned int n,unsigned int p){
 	}
 	}*/
 
-	unsigned int count = (id==0) ? 1 : 0;
-	unsigned int global_count=0;
+	uint64_t count = (id==0) ? 1 : 0;
+	uint64_t global_count=0;
 	for(i = 0; i < size; i++){
 		if (marked[i] == 0) count++;
 	}
-	if (p > 1) MPI_Reduce (&count, &global_count, 1, MPI_UNSIGNED, MPI_SUM,0, MPI_COMM_WORLD);
+	if (p > 1) MPI_Reduce (&count, &global_count, 1, MPI_UNSIGNED_LONG, MPI_SUM,0, MPI_COMM_WORLD);
 
 	elapsed_time += MPI_Wtime();
 	if(!id){
-		printf ("There are {%d} primes less than or equal to %d\n",global_count, n);
-		printf ("SIEVE (%d) %10.6f\n", p, elapsed_time);
+		printf ("There are {%"PRIu64"} primes less than or equal to %"PRIu64"\n",global_count, (uint64_t)n);
+		printf ("SIEVE (%"PRIu64") %10.6f\n", p, elapsed_time);
 	}
 	free(marked);
 	return global_count;
