+++ A03/sieve_original.h
@@ -1,21 +1,24 @@
 #ifndef SIEVE_ORIGINAL
 #define SIEVE_ORIGINAL
 
-#include <math.h>
 
-unsigned int sieve_original(int id, unsigned int n,unsigned int p){
+#include <math.h>
+#include <stdint.h>
+#include <inttypes.h>
+
+uint64_t sieve_original(int id, uint64_t n,uint64_t p){
 	if(id == 0) printf("Executing original sieve\n");
-	unsigned int low_value = 2 + id*(n-1)/p;
-	unsigned int high_value = 1 + (id+1)*(n-1)/p;
-	unsigned int size = high_value - low_value + 1;
+	uint64_t low_value = 2 + ((uint64_t)id)*(n-1)/p;
+	uint64_t  high_value = 1 + ((uint64_t)(id+1))*(n-1)/p;
+	uint64_t size = high_value - low_value + 1;
 
 	MPI_Barrier(MPI_COMM_WORLD);
 	double elapsed_time = -MPI_Wtime();
-	//printf("([%d],%d,%d,%d)\n",id,low_value,high_value,size);
+	//printf("([%d],%"PRIu64",%"PRIu64",%"PRIu64")\n",id,low_value,high_value,size);
 
-	unsigned int proc0_size = (n-1)/p;
+	uint64_t proc0_size = (n-1)/p;
 
-	if ((2 + proc0_size) < (int) sqrt((double) n)) {
+	if ((2 + proc0_size) < (uint64_t) sqrt((double) n)) {
 		if (id == 0) printf ("Too many processes\n");
 		MPI_Finalize();
 		exit (1);
@@ -28,12 +31,12 @@ unsigned int sieve_original(int id, unsigned int n,unsigned int p){
 		exit (1);
 	}
 
-	unsigned int i;
+	uint64_t i;
 	for (i = 0; i < size; i++) marked[i] = 0;
-	unsigned int index;
+	uint64_t index;
 	if (id == 0) index = 0;
-	unsigned int prime = 2;
-	unsigned int first;
+	uint64_t prime = 2;
+	uint64_t first;
 
 	do{
 		if(prime * prime > low_value){
@@ -50,23 +53,23 @@ unsigned int sieve_original(int id, unsigned int n,unsigned int p){
 			while (marked[++index] == 1);
 			prime = index + 2;
 		}
-		if( p > 1 ) MPI_Bcast (&prime,  1, MPI_INT, 0, MPI_COMM_WORLD);
+		if( p > 1 ) MPI_Bcast (&prime,  1, MPI_UNSIGNED_LONG, 0, MPI_COMM_WORLD);
 	}while(prime * prime <= n);
 
 
 
-	unsigned int count = 0;
-	unsigned int global_count;
+	uint64_t count = 0;
+	uint64_t global_count;
 	for(i = 0; i < size; i++){
 		if (marked[i] == 0) count++;
 	}
-	if (p > 1) MPI_Reduce (&count, &global_count, 1, MPI_UNSIGNED, MPI_SUM,0, MPI_COMM_WORLD);
+	if (p > 1) MPI_Reduce (&count, &global_count, 1, MPI_UNSIGNED_LONG, MPI_SUM,0, MPI_COMM_WORLD);
 
 	elapsed_time += MPI_Wtime();
 
 	if(!id){
-		printf ("There are {%d} primes less than or equal to %d\n",global_count, n);
-		printf ("SIEVE (%d) %10.6f\n", p, elapsed_time);
+		printf ("There are {%"PRIu64"} primes less than or equal to %"PRIu64"\n",global_count, (uint64_t)n);
+		printf ("SIEVE (%"PRIu64") %10.6f\n", p, elapsed_time);
 	}
 	free(marked);
 	return global_count;
